Unified font loading code in `roFontRegistry`
